fix: move pagination actions to separate file
